src/app/components/header/header.component.ts
import { CartService } from './../../services/cart.service';
import { Component, Input } from '@angular/core';
import { Cart, CartItem } from 'src/app/models/cart.model';

@Component({
  selector: 'app-header',
  templateUrl: './header.component.html',
})
export class HeaderComponent {
  @Input() productCartQuantity: number = 0;
  @Input() cart: Cart = { items: [] };

  constructor(private cartService: CartService) {}

  getTotal(items: Array<CartItem>): number {
    return this.cartService.getTotal(items);
  }

  onClearCart() {
    this.cartService.clearCart();
  }
}
